reorg in stores props
