+++ src/components/map/map.jsx
@@ -25,10 +25,8 @@ class Map extends React.PureComponent {
   componentDidMount() {
     if (this.mapRef.current) {
       const {offer,
-        //  number
       } = this.props;
 
-      // this.city = offer.number.offerCoord;
       this.city = [52.38333, 4.9];
 
       this.zooms = 12;
