+++ src/reducer.js
@@ -17,10 +17,6 @@ const ActionCreator = ({
     payload: newCity,
   }),
 
-  // 'RESET': () => ({
-  //   type: `RESET`,
-  // }),
-
 });
 
 const reducer = (state = initialState, action) => {
@@ -30,8 +26,6 @@ const reducer = (state = initialState, action) => {
         city: action.payload,
         offerInCity: require(`./mocks/offers-in-${state.city}`).offer,
       });
-    // case `RESET`:
-    //   return Object.assign({}, state, initialState);
   }
 
   return state;
